LayoutTests/webgpu/command-buffers.html

<!DOCTYPE html>
<html>
<script src="js/webgpu-functions.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
promise_test(() => {
    return getBasicDevice().then(function(device) {
        const commandEncoder = device.createCommandEncoder();
        assert_true(commandEncoder instanceof GPUCommandEncoder, "Successfully created GPUCommandEncoder.");
    }, function() {
    });
}, "Create a default GPUCommandEncoder.");

// FIXME: createCommandEncoder should take a GPUCommandEncoderDescriptor, which is currently an empty dictionary.
</script>
</html>
